Fix int to float64

+++ internal/models/models.go
@@ -1,19 +1,19 @@
 package models
 
 type Transaction struct {
-	Id               string `json:"id"`
-	Price            int    `json:"price"`
-	Date             string `json:"date"`
-	Postcode         string `json:"postcode"`
-	TownCity         string `json:"townCity"`
-	Locality         string `json:"locality"`
-	County           string `json:"county"`
-	Street           string `json:"street"`
-	PrimaryAddress   string `json:"primaryAddress"`
-	SecondaryAddress string `json:"secondaryAddress,omitempty"`
-	OldOrNew         string `json:"oldOrNew"`
-	CategoryType     string `json:"categoryType"`
-	PropertyType     string `json:"propertyType"`
-	Duration         string `json:"duration"`
-	RfMonthlyFile    string `json:"rfMonthlyFile"`
+	Id               string  `json:"id"`
+	Price            float64 `json:"price"`
+	Date             string  `json:"date"`
+	Postcode         string  `json:"postcode"`
+	TownCity         string  `json:"townCity"`
+	Locality         string  `json:"locality"`
+	County           string  `json:"county"`
+	Street           string  `json:"street"`
+	PrimaryAddress   string  `json:"primaryAddress"`
+	SecondaryAddress string  `json:"secondaryAddress,omitempty"`
+	OldOrNew         string  `json:"oldOrNew"`
+	CategoryType     string  `json:"categoryType"`
+	PropertyType     string  `json:"propertyType"`
+	Duration         string  `json:"duration"`
+	RfMonthlyFile    string  `json:"rfMonthlyFile"`
 }
